+++ server/src-lib/Hasura/RQL/DDL/Schema/Cache.hs
@@ -42,10 +42,10 @@ import           Hasura.Db
 -- import           Hasura.RQL.DDL.Action
 import           Hasura.RQL.DDL.ComputedField
 -- import           Hasura.RQL.DDL.CustomTypes
+import           Hasura.GraphQL.Schema                    (buildGQLContext)
 import           Hasura.RQL.DDL.Deps
 import           Hasura.RQL.DDL.EventTrigger
 import           Hasura.RQL.DDL.RemoteSchema
-import           Hasura.GraphQL.Schema                    (buildGQLContext)
 import           Hasura.RQL.DDL.Schema.Cache.Common
 import           Hasura.RQL.DDL.Schema.Cache.Dependencies
 import           Hasura.RQL.DDL.Schema.Cache.Fields
@@ -203,7 +203,7 @@ buildSchemaCacheRule = proc (catalogMetadata, invalidationKeys) -> do
     buildAndCollectInfo = proc (catalogMetadata, invalidationKeys) -> do
       let CatalogMetadata tables relationships permissions
             eventTriggers remoteSchemas functions allowlistDefs
-            computedFields customTypes actions = catalogMetadata
+            computedFields catalogCustomTypes actions = catalogMetadata
 
       -- tables
       tableRawInfos <- buildTableCache -< (tables, Inc.selectD #_ikMetadata invalidationKeys)
@@ -261,26 +261,22 @@ buildSchemaCacheRule = proc (catalogMetadata, invalidationKeys) -> do
             & HS.fromList
 
       -- custom types
-      -- resolvedCustomTypes <- bindA -< resolveCustomTypes tableCache customTypes
+      -- let CatalogCustomTypes customTypes pgScalars = catalogCustomTypes
+      -- maybeResolvedCustomTypes <-
+      --   (| withRecordInconsistency
+      --        (bindErrorA -< resolveCustomTypes tableCache customTypes pgScalars)
+      --    |) (MetadataObject MOCustomTypes $ toJSON customTypes)
 
-      -- actions
-      -- actionCache <- (mapFromL _amName actions >- returnA)
-      --   >-> (| Inc.keyed (\_ action -> do
-      --          let ActionMetadata name comment def actionPermissions = action
-      --              metadataObj = MetadataObject (MOAction name) $ toJSON $
-      --                            CreateAction name def comment
-      --              addActionContext e = "in action " <> name <<> "; " <> e
-      --          (| withRecordInconsistency (
-      --             (| modifyErrA ( do
-      --                  (resolvedDef, outFields) <- bindErrorA -< resolveAction resolvedCustomTypes def
-      --                  let permissionInfos = map (ActionPermissionInfo . _apmRole) actionPermissions
-      --                      permissionMap = mapFromL _apiRole permissionInfos
-      --                  returnA -< ActionInfo name outFields resolvedDef permissionMap comment
-      --                )
-      --              |) addActionContext)
-      --           |) metadataObj)
-      --        |)
-      --   >-> (\actionMap -> returnA -< M.catMaybes actionMap)
+      -- -- actions
+      -- actionCache <- case maybeResolvedCustomTypes of
+      --   Just resolvedCustomTypes -> buildActions -< ((resolvedCustomTypes, pgScalars), actions)
+
+      --   -- If the custom types themselves are inconsistent, we can’t really do
+      --   -- anything with actions, so just mark them all inconsistent.
+      --   Nothing -> do
+      --     recordInconsistencies -< ( map mkActionMetadataObject actions
+      --                              , "custom types are inconsistent" )
+      --     returnA -< M.empty
 
       -- remote schemas
       let remoteSchemaInvalidationKeys = Inc.selectD #_ikRemoteSchemas invalidationKeys
@@ -293,12 +289,18 @@ buildSchemaCacheRule = proc (catalogMetadata, invalidationKeys) -> do
         , _boRemoteSchemas = remoteSchemaMap
         , _boAllowlist = allowList
         -- , _boCustomTypes = resolvedCustomTypes
+        -- If 'maybeResolvedCustomTypes' is 'Nothing', then custom types are inconsinstent.
+        -- In such case, use empty resolved value of custom types.
+        -- , _boCustomTypes = fromMaybe (NonObjectTypeMap mempty, mempty) maybeResolvedCustomTypes
         }
 
     mkEventTriggerMetadataObject (CatalogEventTrigger qt trn configuration) =
       let objectId = MOTableObj qt $ MTOTrigger trn
           definition = object ["table" .= qt, "configuration" .= configuration]
       in MetadataObject objectId definition
+
+    mkActionMetadataObject (ActionMetadata name comment defn _) =
+      MetadataObject (MOAction name) (toJSON $ CreateAction name defn comment)
 
     mkRemoteSchemaMetadataObject remoteSchema =
       MetadataObject (MORemoteSchema (_arsqName remoteSchema)) (toJSON remoteSchema)
@@ -375,6 +377,28 @@ buildSchemaCacheRule = proc (catalogMetadata, invalidationKeys) -> do
                runExceptT $ addRemoteSchemaP2Setup remoteSchema)
            |) (mkRemoteSchemaMetadataObject remoteSchema)
 
+    -- buildActions
+    --   :: ( ArrowChoice arr, Inc.ArrowDistribute arr, Inc.ArrowCache m arr
+    --      , ArrowWriter (Seq CollectedInfo) arr, MonadIO m )
+    --   => ( ((NonObjectTypeMap, AnnotatedObjects), HashSet PGScalarType)
+    --      , [ActionMetadata]
+    --      ) `arr` HashMap ActionName ActionInfo
+    -- buildActions = buildInfoMap _amName mkActionMetadataObject buildAction
+    --   where
+    --     buildAction = proc ((resolvedCustomTypes, pgScalars), action) -> do
+    --       let ActionMetadata name comment def actionPermissions = action
+    --           addActionContext e = "in action " <> name <<> "; " <> e
+    --       (| withRecordInconsistency (
+    --          (| modifyErrA (do
+    --               (resolvedDef, outObject, reusedPgScalars) <- liftEitherA <<< bindA -<
+    --                 runExceptT $ resolveAction resolvedCustomTypes pgScalars def
+    --               let permissionInfos = map (ActionPermissionInfo . _apmRole) actionPermissions
+    --                   permissionMap = mapFromL _apiRole permissionInfos
+    --               returnA -< ActionInfo name outObject resolvedDef permissionMap reusedPgScalars comment)
+    --           |) addActionContext)
+    --        |) (mkActionMetadataObject action)
+
+
     -- -- Builds the GraphQL schema and merges in remote schemas. This function is kind of gross, as
     -- -- it’s possible for the remote schema merging to fail, at which point we have to mark them
     -- -- inconsistent. This means we have to accumulate the consistent remote schemas as we go, in
@@ -389,7 +413,7 @@ buildSchemaCacheRule = proc (catalogMetadata, invalidationKeys) -> do
     --      , ActionCache
     --      ) `arr` (RemoteSchemaMap, GS.GCtxMap, GS.GCtx)
     -- buildGQLSchema = proc (tableCache, functionCache, remoteSchemas, customTypes, actionCache) -> do
-    --   baseGQLSchema <- bindA -< GS.mkGCtxMap (snd customTypes) tableCache functionCache actionCache
+    --   baseGQLSchema <- bindA -< GS.mkGCtxMap tableCache functionCache actionCache
     --   (| foldlA' (\(remoteSchemaMap, gqlSchemas, remoteGQLSchemas)
     --                (remoteSchemaName, (remoteSchema, metadataObject)) ->
     --        (| withRecordInconsistency (do